requirements.txt
cycler==0.10.0
Django==1.11.6
djangorestframework==3.7.1
nltk==3.2.5
numpy==1.13.3
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
scikit-learn==0.19.1
six==1.11.0
